merge
merged the branch, created a superclass GroundEnemy <feature, bug
