+++ src/main/java/sem/group47/entity/enemies/GroundEnemy.java
@@ -66,8 +66,9 @@ public class GroundEnemy extends Enemy {
 		}
 		if (isCaught()) {
 			setDy(getDy() - getFloatSpeed());
-			if (getDy() < getMaxFloatSpeed())
+			if (getDy() < getMaxFloatSpeed()) {
 			 setDy(getMaxFloatSpeed());
+			}
 			setDx(0);
 		} else if (isFalling()) {
 			setDy(getDy() + getFallSpeed());
@@ -115,6 +116,15 @@ public class GroundEnemy extends Enemy {
 		getNextYPosition();
 		checkTileMapCollision();
 		setPosition(getXposNew(), getYposNew());
+		updateStates();
 	}
-	
+
+	public final void updateStates() {
+		if(isCaught() && (System.nanoTime() - getTimeCaught())/1000000000.0d > getTimeUntillBreakFree()) {
+			setCaught(false);
+		}
+		if(isAngry() && (System.nanoTime() - getAngryTime())/1000000000.0d > 10) {
+			setIsAngry(false);
+		}
+	}
 }
